Switched Compiler to clang-llvm Added 64bit architecture Fixed deprecation warnings Fixed possible leaks from static analzyer run Replaced menu bar icon with one that has transparent corners

+++ MKConsoleWindow.m
@@ -45,7 +45,7 @@ extern OSStatus CGSClearWindowTags(const CGSConnection cid, const CGSWindow wid,
     @implementation MKConsoleWindow
 //---------------------------------------------------------------------------------------
 
-- (id)initWithContentRect:(NSRect)contentRect styleMask:(unsigned int)styleMask backing:(NSBackingStoreType)backingType defer:(BOOL)flag
+- (id)initWithContentRect:(NSRect)contentRect styleMask:(NSUInteger)styleMask backing:(NSBackingStoreType)backingType defer:(BOOL)flag
 {
     self = [super initWithContentRect:contentRect styleMask:NSBorderlessWindowMask backing:NSBackingStoreBuffered defer:NO];
     NSAssert1(self != nil, @"%s Failed to create window instance", __PRETTY_FUNCTION__);
